+++ Final1096917/Final1096917/Lavadora.h
@@ -7,7 +7,7 @@ using namespace std;
 class Lavadora: public Electrodomestico
 {
 public:
-	Lavadora(int ePrecio, string eColor, string eConsumo, int ePeso, int lCarga);
+	Lavadora(float ePrecio, string eColor, string eConsumo, int ePeso, int lCarga);
 	int getCarga();
 	void setCarga(int lCarga);
 	void toString();
